Added Editor Only Functionality
Added a separate editor only object as a sub-class of the MCURegistry definition. This is to group the functionality together and add some new usability.

Added a menu option to remove the adapter scripting defines from the project and forcing a regeneration with the defined symbols.

Also added a simple function to search for directories with a matching structure to allow for relocation of editor generated items without requiring concrete directory locations.